Ship Combat
Enemy ship status check added
refactoring shipmapprops beginning

+++ Junkstar Rework/Assets/Scripts/Props/ShipWeaponProps.cs
@@ -9,7 +9,7 @@ public class ShipWeaponProps : MonoBehaviour
     public GameObject projectile;
     public string weaponName;
     public string weaponTxt;
-    public float accuracy;
+    public int accuracy;
     public float speed;
     public int cooldown;
     public int curCooldown;
